Session: only manage the cookie, not associated data:
This is because in order to access associated data, the consumer of the library
would still need to store an Arc to this plugin in their shared state. If they
are doing that, they could easily create a CHashMap there themselves, and there
would be less complication for them.  They also now have the flexibility to use
any other state storage mechanism they choose.